refactor(homecards): adjust component declaration
