+++ FinEdgeFrontend/src/Components/HomePage/DashboardCard.tsx
@@ -1,6 +1,11 @@
 import { Card, CardBody } from "@progress/kendo-react-all";
 
-const DashboardCard: React.FC<{children: React.ReactNode; className?: string;}> = ({ children, className = "" }) => (
+type DashboardCardProps = { 
+  children: React.ReactNode; 
+  className?: string;
+}
+
+const DashboardCard = ({ children, className }: DashboardCardProps)  => (
   <div className={`bg-white rounded-lg shadow ${className}`}>
     <Card className="border-none shadow-none h-full">
       <CardBody>{children}</CardBody>
